Sredjivanje nekih testova
