+++ user-service/src/test/java/com/banka1/user/SearchIntegrationTest.java
@@ -20,7 +20,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ActiveProfiles("test")
-public class PregledIntegrationTest {
+public class SearchIntegrationTest {
     @LocalServerPort
     private int port;
