+++ user-service/src/test/java/com/banka1/user/controllers/CustomerControllerTest.java
@@ -3,6 +3,7 @@ package com.banka1.user.controllers;
 
 import com.banka1.user.DTO.request.CreateCustomerRequest;
 import com.banka1.user.DTO.request.UpdateCustomerRequest;
+import com.banka1.user.DTO.response.CustomerResponse;
 import com.banka1.user.mapper.CustomerMapper;
 import com.banka1.user.model.Customer;
 import com.banka1.user.model.helper.Gender;
@@ -13,9 +14,13 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.springframework.http.MediaType;
+import org.springframework.test.util.AssertionErrors;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 import java.util.List;
@@ -146,5 +151,60 @@ class CustomerControllerTest {
                 .andExpect(status().isNotFound())
                 .andExpect(jsonPath("$.success").value(false))
                 .andExpect(jsonPath("$.error").value("Korisnik nije pronađen"));
+    }
+
+    @Test
+    void findByIdSuccess() throws Exception {
+        String id = "1";
+        var response = new CustomerResponse(
+                1L,
+                "Petar",
+                "Petrovic",
+                "ppetrovic",
+                1234567890L,
+                Gender.MALE,
+                "ppetrovic@banka.rs",
+                "99999999",
+                "Ulica",
+                List.of());
+
+        Mockito.when(customerService.findById(id)).thenReturn(response);
+
+        var responseContent = mockMvc.perform(MockMvcRequestBuilders.get("/api/customer/" + id))
+                .andExpect(MockMvcResultMatchers.status().isOk())
+                .andExpect(MockMvcResultMatchers.jsonPath("$.success").value(true))
+                .andExpect(MockMvcResultMatchers.jsonPath("$.data.id").value(id))
+                .andReturn().getResponse().getContentAsString();
+        var responseJson = objectMapper.convertValue(objectMapper.readTree(responseContent).get("data"), CustomerResponse.class);
+
+        AssertionErrors.assertEquals("Response", response, responseJson);
+
+        Mockito.verify(customerService).findById(id);
+    }
+
+    @Test
+    void findByIdNotFound() throws Exception {
+        String id = "1";
+
+        Mockito.when(customerService.findById(id)).thenReturn(null);
+
+        mockMvc.perform(MockMvcRequestBuilders.get("/api/customer/" + id))
+                .andExpect(MockMvcResultMatchers.status().isNotFound())
+                .andExpect(MockMvcResultMatchers.jsonPath("$.success").value(false));
+
+        Mockito.verify(customerService).findById(id);
+    }
+
+    @Test
+    void findByIdBadRequest() throws Exception {
+        String id = "1";
+
+        Mockito.when(customerService.findById(id)).thenThrow(new RuntimeException("Poruka"));
+
+        mockMvc.perform(MockMvcRequestBuilders.get("/api/customer/" + id))
+                .andExpect(MockMvcResultMatchers.status().isBadRequest())
+                .andExpect(MockMvcResultMatchers.jsonPath("$.success").value(false));
+
+        Mockito.verify(customerService).findById(id);
     }
 }
