+++ user-service/src/test/java/com/banka1/user/controllers/EmployeeControllerTest.java
@@ -3,8 +3,12 @@ package com.banka1.user.controllers;
 import com.banka1.user.DTO.request.CreateEmployeeRequest;
 import com.banka1.user.DTO.request.UpdateEmployeeRequest;
 import com.banka1.user.DTO.request.UpdatePermissionsRequest;
+import com.banka1.user.DTO.response.EmployeeResponse;
 import com.banka1.user.model.Employee;
+import com.banka1.user.model.helper.Department;
+import com.banka1.user.model.helper.Gender;
 import com.banka1.user.model.helper.Permission;
+import com.banka1.user.model.helper.Position;
 import com.banka1.user.service.EmployeeService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
@@ -15,7 +19,10 @@ import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
+import org.springframework.test.util.AssertionErrors;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 import java.util.List;
@@ -118,5 +125,65 @@ class EmployeeControllerTest {
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.success").value(true))
                 .andExpect(jsonPath("$.data").value("Permisije korisnika ažurirane"));
+    }
+
+    @Test
+    void findByIdSuccess() throws Exception {
+        String id = "1";
+        var response = new EmployeeResponse(
+                1L,
+                "Petar",
+                "Petrovic",
+                "ppetrovic",
+                "9999-09-09",
+                Gender.MALE,
+                "ppetrovic@banka.rs",
+                "99999999",
+                "Ulica",
+                Position.WORKER,
+                Department.IT,
+                true,
+                false,
+                List.of());
+
+        Mockito.when(employeeService.findById(id)).thenReturn(response);
+
+        var responseContent = mockMvc.perform(MockMvcRequestBuilders.get("/api/users/employees/" + id))
+                .andExpect(MockMvcResultMatchers.status().isOk())
+                .andExpect(MockMvcResultMatchers.jsonPath("$.success").value(true))
+                .andExpect(MockMvcResultMatchers.jsonPath("$.data.id").value(id))
+                .andReturn().getResponse().getContentAsString();
+        System.out.println(responseContent);
+        var responseJson = objectMapper.convertValue(objectMapper.readTree(responseContent).get("data"), EmployeeResponse.class);
+
+        AssertionErrors.assertEquals("Response", response, responseJson);
+
+        Mockito.verify(employeeService).findById(id);
+    }
+
+    @Test
+    void findByIdNotFound() throws Exception {
+        String id = "1";
+
+        Mockito.when(employeeService.findById(id)).thenReturn(null);
+
+        mockMvc.perform(MockMvcRequestBuilders.get("/api/users/employees/" + id))
+                .andExpect(MockMvcResultMatchers.status().isNotFound())
+                .andExpect(MockMvcResultMatchers.jsonPath("$.success").value(false));
+
+        Mockito.verify(employeeService).findById(id);
+    }
+
+    @Test
+    void findByIdBadRequest() throws Exception {
+        String id = "1";
+
+        Mockito.when(employeeService.findById(id)).thenThrow(new RuntimeException("Poruka"));
+
+        mockMvc.perform(MockMvcRequestBuilders.get("/api/users/employees/" + id))
+                .andExpect(MockMvcResultMatchers.status().isBadRequest())
+                .andExpect(MockMvcResultMatchers.jsonPath("$.success").value(false));
+
+        Mockito.verify(employeeService).findById(id);
     }
 }
