+++ user-service/src/test/java/com/banka1/user/services/EmployeeServiceTest.java
@@ -3,10 +3,14 @@ package com.banka1.user.services;
 import com.banka1.user.DTO.request.CreateEmployeeRequest;
 import com.banka1.user.DTO.request.UpdateEmployeeRequest;
 import com.banka1.user.DTO.request.UpdatePermissionsRequest;
+import com.banka1.user.DTO.response.EmployeeResponse;
+import com.banka1.user.DTO.response.EmployeesPageResponse;
 import com.banka1.user.listener.MessageHelper;
 import com.banka1.user.model.Employee;
+import com.banka1.user.model.helper.Department;
 import com.banka1.user.model.helper.Gender;
 import com.banka1.user.model.helper.Permission;
+import com.banka1.user.model.helper.Position;
 import com.banka1.user.repository.EmployeeRepository;
 import com.banka1.user.service.EmployeeService;
 import org.junit.jupiter.api.BeforeEach;
@@ -14,12 +18,17 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.jms.core.JmsTemplate;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.test.util.AssertionErrors;
 
 import java.util.List;
 import java.util.Optional;
@@ -145,5 +154,104 @@ class EmployeeServiceTest {
         when(employeeRepository.findById(1L)).thenReturn(Optional.empty());
 
         assertThrows(RuntimeException.class, () -> employeeService.updateEmployee(1L, dto));
+    }
+
+    @Test
+    void findByIdSuccess() {
+        String id = "1";
+        var entity = getEmployee();
+
+        Mockito.when(employeeRepository.findById(1L)).thenReturn(Optional.of(entity));
+
+        var response = getEmployeeResponse();
+
+        AssertionErrors.assertEquals("Response", response, employeeService.findById(id));
+
+        Mockito.verify(employeeRepository).findById(1L);
+    }
+
+    private static EmployeeResponse getEmployeeResponse() {
+        return new EmployeeResponse(
+                1L,
+                "Petar",
+                "Petrovic",
+                "ppetrovic",
+                "9999-09-09",
+                Gender.MALE,
+                "ppetrovic@banka.rs",
+                "99999999",
+                "Ulica",
+                Position.WORKER,
+                Department.IT,
+                true,
+                false,
+                List.of());
+    }
+
+    private static Employee getEmployee() {
+        var entity = new Employee();
+
+        entity.setFirstName("Petar");
+        entity.setLastName("Petrovic");
+        entity.setUsername("ppetrovic");
+        entity.setBirthDate("9999-09-09");
+        entity.setGender(Gender.MALE);
+        entity.setEmail("ppetrovic@banka.rs");
+        entity.setPhoneNumber("99999999");
+        entity.setAddress("Ulica");
+        entity.setPosition(Position.WORKER);
+        entity.setDepartment(Department.IT);
+        entity.setActive(true);
+        entity.setIsAdmin(false);
+        entity.setPermissions(List.of());
+        entity.setId(1L);
+        return entity;
+    }
+
+    @Test
+    void findByIdNotFound() {
+        String id = "1";
+
+        Mockito.when(employeeRepository.findById(1L)).thenReturn(Optional.empty());
+
+        AssertionErrors.assertNull("Response", employeeService.findById(id));
+
+        Mockito.verify(employeeRepository).findById(1L);
+    }
+
+    @Test
+    void findByIdInvalidId() {
+        String id = "Petar";
+
+        try {
+            employeeService.findById(id);
+            fail("No exception.");
+        } catch (Exception e) {
+            AssertionErrors.assertNotNull("Error", e);
+        }
+    }
+
+    @Test
+    void search() {
+        var response = new EmployeesPageResponse(1, List.of(getEmployeeResponse()));
+
+        var pageable = PageRequest.of(0, 10, Sort.by(Sort.Direction.ASC, "id"));
+        Mockito.when(employeeRepository.findAll(pageable))
+                .thenReturn(new PageImpl<>(List.of(getEmployee()), pageable, 1));
+
+        AssertionErrors.assertEquals("Response", response,
+                employeeService.search(0, 10, Optional.empty(), Optional.empty(), Optional.empty(), Optional.empty()));
+
+        Mockito.verify(employeeRepository).findAll(pageable);
+    }
+
+    @Test
+    void searchNoFilterValue() {
+        try {
+            employeeService.search(0, 10, Optional.empty(), Optional.empty(), Optional.of("firstName"), Optional.empty());
+            fail("No exception.");
+        } catch (Exception e) {
+            AssertionErrors.assertNotNull("Error", e);
+        }
     }
 }
